Only memoize full entities.
This is necessary for direct left-recursion as with every single atom being memoized,
even what appears to be a direct left-recursion actually becomes an indirect left-recursion.

I also suspect that memoizing even the smallest fragments may cause significant memory growth and
object creation, both of which can be a problem in ruby in terms of causing degredation of other
performance metrics, anyways.